Moved test files out of test folder, tests seem to pass now. Will explore later
