+++ src/Card.test.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import Card from '../Card.js';
+import Card from './Card.js';
 import { shallow } from 'enzyme';
 
 let id = "1";
