+++ src/css/CardContainer.scss
@@ -1,0 +1,5 @@
+.card-container {
+  display: grid;
+  grid-template-columns: 1fr 1fr 1fr 1fr; 
+}
+
