+++ src/css/ParlorArea.scss
@@ -1,0 +1,3 @@
+.parlor-area {
+  
+}
